Moved classes around backend to better suit onion architecture
- moved DTO classes from `Infrastructure/DTO` to `Core/DTO`
- moved some detached response types from `Core/Services` to `Core/DTO`
- moved controller message types from `Infrastructure/Commands` to `WebAPI/Model`
