+++ CloudDrive.Core/Services/IFileManagerService.cs
@@ -1,9 +1,11 @@
-﻿namespace CloudDrive.Core.Services
+﻿using CloudDrive.Core.DTO;
+
+namespace CloudDrive.Core.Services
 {
     public interface IFileManagerService
     {
-        Task<CreateFileResult> CreateFile(Guid userId, Stream inputStream, string fileName, string clientDirPath);
-        Task<GetFileResult?> GetFileVersion(Guid fileId, int versionNr);
-        Task<GetFileResult?> GetLatestFileVersion(Guid fileId);
+        Task<CreateFileResultDTO> CreateFile(Guid userId, Stream inputStream, string fileName, string clientDirPath);
+        Task<GetFileResultDTO?> GetFileVersion(Guid fileId, int versionNr);
+        Task<GetFileResultDTO?> GetLatestFileVersion(Guid fileId);
     }
 }
