--- CloudDrive.Infrastructure/Commands/SignUpRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace CloudDrive.Infrastructure.Commands
-{
-	public class SignUpRequest
-	{
-		public string Email { get; set; }
-		public string Password { get; set; }
-	}
-}
