--- CloudDrive.Infrastructure/Commands/SyncGetResponse.cs
@@ -1,9 +1,0 @@
-﻿using CloudDrive.Infrastructure.DTO;
-
-namespace CloudDrive.Infrastructure.Commands
-{
-    public class SyncGetResponse
-    {
-        public required FileVersionDTO[] CurrentFileVersionsInfos {  get; set; }
-    }
-}
