+++ CloudDrive.WebAPI/Controllers/AuthController.cs
@@ -1,5 +1,5 @@
 ﻿using CloudDrive.Core.Services;
-using CloudDrive.Infrastructure.Commands;
+using CloudDrive.WebAPI.Model;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CloudDrive.WebAPI.Controllers
