+++ CloudDrive.WebAPI/Controllers/SyncController.cs
@@ -1,7 +1,7 @@
 ﻿using CloudDrive.Core.Services;
-using CloudDrive.Infrastructure.Commands;
 using CloudDrive.Infrastructure.DTO;
 using CloudDrive.WebAPI.Extensions;
+using CloudDrive.WebAPI.Model;
 using Mapster;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
