+++ CloudDrive.WebAPI/Model/SignInRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace CloudDrive.WebAPI.Model
+{
+    public class SignInRequest
+    {
+        public string Email { get; set; }
+        public string Password { get; set; }
+    }
+}
