+++ CloudDrive.WebAPI/Model/SignUpRequest.cs
@@ -1,6 +1,6 @@
-﻿namespace CloudDrive.Infrastructure.Commands
+﻿namespace CloudDrive.WebAPI.Model
 {
-    public class SignInRequest
+    public class SignUpRequest
     {
         public string Email { get; set; }
         public string Password { get; set; }
