+++ CloudDrive.WebAPI/Model/SyncGetResponse.cs
@@ -1,0 +1,9 @@
+﻿using CloudDrive.Infrastructure.DTO;
+
+namespace CloudDrive.WebAPI.Model
+{
+    public class SyncGetResponse
+    {
+        public required FileVersionDTO[] CurrentFileVersionsInfos { get; set; }
+    }
+}
